use of LfAsymLine class
