+++ src/main/java/com/powsybl/openloadflow/ac/equations/asym/AbstractAsymmetricalBranchFlowEquationTerm.java
@@ -11,9 +11,9 @@ package com.powsybl.openloadflow.ac.equations.asym;
 import com.powsybl.openloadflow.ac.equations.AcEquationType;
 import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.equations.AbstractElementEquationTerm;
-import com.powsybl.openloadflow.network.extensions.AsymBranch;
+import com.powsybl.openloadflow.network.LfAsymLine;
+import com.powsybl.openloadflow.network.LfAsymLineAdmittanceMatrix;
 import com.powsybl.openloadflow.network.LfBranch;
-import com.powsybl.openloadflow.network.extensions.AsymLineAdmittanceMatrix;
 
 /**
  * @author Geoffroy Jamgotchian {@literal <geoffroy.jamgotchian at gmail.com>}
@@ -22,11 +22,11 @@ import com.powsybl.openloadflow.network.extensions.AsymLineAdmittanceMatrix;
 abstract class AbstractAsymmetricalBranchFlowEquationTerm extends AbstractElementEquationTerm<LfBranch, AcVariableType, AcEquationType> {
 
     // Classical line parameters are replaced by a 12x12 admittance matrix
-    protected final AsymLineAdmittanceMatrix y;
+    protected final LfAsymLineAdmittanceMatrix y;
 
     protected AbstractAsymmetricalBranchFlowEquationTerm(LfBranch branch) {
         super(branch);
-        AsymBranch asymLine = (AsymBranch) branch.getProperty(AsymBranch.PROPERTY_ASYMMETRICAL);
+        LfAsymLine asymLine = branch.getAsymLine();
         if (asymLine == null) {
             throw new IllegalStateException("Line : " + branch.getId() + " has no asymmetric extension but is required here ");
         }
